Create data_hathaway
USS Hathaway form Wave 20, ship 1 of 3

I made a tweek to the schema for id numbers as I saw you were using 71 to denote ship sets not part of a OP I kept that but added a new way to categorize. for any new ships I will likely be using "71" then the wave number "20" then the number for order that ship was made in the case "1"

So the Hathaway is 71201 and each "id" ends with that number instead of using the ships name.

I will be adding the Scorpion 4 and Halik soon. I am also almost done with wave 21 and after that will move on the the missing resources and prize ships. I am really hoping you get this message and I look forward to working together.